move filter project into project filter file
