+++ optd-datafusion-repr/src/rules/project_transpose.rs
@@ -2,4 +2,3 @@ pub mod project_transpose_common;
 pub mod project_merge;
 pub mod project_filter_transpose;
 pub mod project_join_transpose;
-pub mod filter_project_transpose;
