+++ optd-datafusion-repr/src/rules/project_transpose/project_filter_transpose.rs
@@ -1,4 +1,5 @@
 use std::collections::HashMap;
+use std::vec;
 
 use optd_core::rules::{Rule, RuleMatcher};
 use optd_core::{optimizer::Optimizer, rel_node::RelNode};
@@ -15,9 +16,6 @@ fn merge_exprs(first: ExprList, second: ExprList) -> ExprList {
     ExprList::new(res_vec)
 }
 
-// pushes projections through filters
-// adds a projection node after a filter node 
-// only keeping necessary columns (proj node exprs + filter col exprs)
 define_rule!(
     ProjectFilterTransposeRule,
     apply_projection_filter_transpose,
@@ -28,6 +26,9 @@ define_rule!(
     )
 );
 
+/// pushes projections through filters
+/// adds a projection node after a filter node 
+/// only keeping necessary columns (proj node exprs + filter col exprs)
 fn apply_projection_filter_transpose(
     _optimizer: &impl Optimizer<OptRelNodeTyp>,
     ProjectFilterTransposeRulePicks { child, cond, exprs }: ProjectFilterTransposeRulePicks,
@@ -71,6 +72,32 @@ fn apply_projection_filter_transpose(
     vec![top_proj_node.into_rel_node().as_ref().clone()]
 }
 
+define_rule!(
+    FilterProjectTransposeRule,
+    apply_filter_project_transpose,
+    (Filter, (Projection, child, [exprs]), [cond])
+);
+
+/// Datafusion only pushes filter past project when the project does not contain
+/// volatile (i.e. non-deterministic) expressions that are present in the filter
+/// Calcite only checks if the projection contains a windowing calculation
+/// We check neither of those things and do it always (which may be wrong)
+fn apply_filter_project_transpose(
+    _optimizer: &impl Optimizer<OptRelNodeTyp>,
+    FilterProjectTransposeRulePicks { child, exprs, cond }: FilterProjectTransposeRulePicks,
+) -> Vec<RelNode<OptRelNodeTyp>> {
+    let child = PlanNode::from_group(child.into());
+    let cond_as_expr = Expr::from_rel_node(cond.into()).unwrap();
+    let exprs = ExprList::from_rel_node(exprs.into()).unwrap();
+
+    let proj_col_map = ProjectionMapping::build(&exprs).unwrap();
+    let rewritten_cond = proj_col_map.rewrite_filter_cond(cond_as_expr.clone(), false);
+
+    let new_filter_node = LogicalFilter::new(child, rewritten_cond);
+    let new_proj = LogicalProjection::new(new_filter_node.into_plan_node(), exprs);
+    vec![new_proj.into_rel_node().as_ref().clone()]
+}
+
 #[cfg(test)]
 mod tests {
     use std::sync::Arc;
@@ -82,10 +109,11 @@ mod tests {
             BinOpExpr, BinOpType, ColumnRefExpr, ConstantExpr, ExprList, LogOpExpr, LogOpType, 
             LogicalFilter, LogicalProjection, LogicalScan, OptRelNode, OptRelNodeTyp
         },
-        rules::ProjectFilterTransposeRule,
+        rules::{ProjectFilterTransposeRule, FilterProjectTransposeRule},
         testing::new_test_optimizer,
     };
 
+    // ProjectFilterTransposeRule Tests
     #[test]
     fn push_proj_past_filter_basic_1() {
         // convert proj -> filter -> scan to filter -> proj -> scan
@@ -332,5 +360,84 @@ mod tests {
 
         assert!(matches!(plan.child(0).child(0).typ, OptRelNodeTyp::Projection));
         assert_eq!(plan.child(0).child(0).child(1), bot_proj_exprs);
+    }
+
+    // FilterProjectTransposeRule Tests
+    #[test]
+    fn push_filter_past_proj_basic() {
+        let mut test_optimizer = new_test_optimizer(Arc::new(FilterProjectTransposeRule::new()));
+
+        let scan = LogicalScan::new("customer".into());
+        let proj = LogicalProjection::new(scan.into_plan_node(), ExprList::new(vec![ColumnRefExpr::new(0).into_expr()]));
+
+        let filter_expr = BinOpExpr::new(
+            ColumnRefExpr::new(0).into_expr(),
+            ConstantExpr::int32(5).into_expr(),
+            BinOpType::Eq,
+        )
+        .into_expr();
+
+        let filter = LogicalFilter::new(proj.into_plan_node(), filter_expr);
+        let plan = test_optimizer.optimize(filter.into_rel_node()).unwrap();
+
+        assert_eq!(plan.typ, OptRelNodeTyp::Projection);
+        assert!(matches!(plan.child(0).typ, OptRelNodeTyp::Filter));
+    }
+
+    #[test]
+    fn push_filter_past_proj_adv() {
+        let mut test_optimizer = new_test_optimizer(Arc::new(FilterProjectTransposeRule::new()));
+
+        let scan = LogicalScan::new("customer".into());
+        let proj = LogicalProjection::new(
+            scan.into_plan_node(),
+            ExprList::new(vec![
+                ColumnRefExpr::new(0).into_expr(),
+                ColumnRefExpr::new(4).into_expr(),
+                ColumnRefExpr::new(5).into_expr(),
+                ColumnRefExpr::new(7).into_expr(),
+            ]),
+        );
+
+        let filter_expr = LogOpExpr::new(
+            LogOpType::And,
+            ExprList::new(vec![
+                BinOpExpr::new(
+                    // This one should be pushed to the left child
+                    ColumnRefExpr::new(1).into_expr(),
+                    ConstantExpr::int32(5).into_expr(),
+                    BinOpType::Eq,
+                )
+                .into_expr(),
+                BinOpExpr::new(
+                    // This one should be pushed to the right child
+                    ColumnRefExpr::new(3).into_expr(),
+                    ConstantExpr::int32(6).into_expr(),
+                    BinOpType::Eq,
+                )
+                .into_expr(),
+            ]),
+        );
+
+        let filter = LogicalFilter::new(proj.into_plan_node(), filter_expr.into_expr());
+
+        let plan = test_optimizer.optimize(filter.into_rel_node()).unwrap();
+
+        assert!(matches!(plan.typ, OptRelNodeTyp::Projection));
+        let plan_filter = LogicalFilter::from_rel_node(plan.child(0)).unwrap();
+        assert!(matches!(plan_filter.0.typ(), OptRelNodeTyp::Filter));
+        let plan_filter_expr =
+            LogOpExpr::from_rel_node(plan_filter.cond().into_rel_node()).unwrap();
+        assert!(matches!(plan_filter_expr.op_type(), LogOpType::And));
+        let op_0 = BinOpExpr::from_rel_node(plan_filter_expr.children()[0].clone().into_rel_node())
+            .unwrap();
+        let col_0 =
+            ColumnRefExpr::from_rel_node(op_0.left_child().clone().into_rel_node()).unwrap();
+        assert_eq!(col_0.index(), 4);
+        let op_1 = BinOpExpr::from_rel_node(plan_filter_expr.children()[1].clone().into_rel_node())
+            .unwrap();
+        let col_1 =
+            ColumnRefExpr::from_rel_node(op_1.left_child().clone().into_rel_node()).unwrap();
+        assert_eq!(col_1.index(), 7);
     }
 }
